Add a simple benchmark to check FP capabilities
Results for 1000 iterations:
- BID+USB: 3667ms   (bid128)
- BID+Bat: 10091ms
- DBL+USB: 668ms    (double)
- DBL+Bat: 1963ms
- FPU+USB: 43ms     (float)
- FPU+Bat: 96ms

So we have close to two orders of magnitude between hardware FP and BID,
and a factor of 15 between double and float.

For 1M iterations on simulator:
- BID: 9847ms
- FPU: 172ms
- DBL: 240ms

So here, the ratio is ~50. And the simulator is roughly 370x faster than the
calculator, which we will need to remember as we implement complicated
or expensive algorithms.

Also, we will need to really prefer float wherever possible, since this is
processed in hardware, whereas I suspect that double is only software on that
CPU.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>