Clear the command when we clear errors
This avoids having the previous command show up in case the new error does not
have a command. An example of scenario is hitting backspace (Drop), getting a
drop error, then trying to get help on a key that has no help yet. The message
will be something like "Drop error: no help for +", which is weird.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>