Another pass at font fine-tuning
Reset the vertical font positions in the table to what they were initially,
add -y option to ttf2font to perform the adjustment in code.

This preserves the relative positions of lowercase 'a' and 'g'  much better,
and allows the guides in the font editor to show at the right position.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>