Do not call object methods from the test thread
The render() method may cause garbage collection, and that was never written to
be thread-safe (and probably never will for performance reasons on the target of
choice). Instead, add instrumentation in the stack rendering code to save a copy
of the stack top-level information that the test suite can use.

With that, we stop seeing instant data corruption each time the tests run.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>