Correct the digit rendering order when rendering to file
The integer and bignum classes were relying on a rather nasty workaround where
they render digits starting with the last one (this is because that's what
modulo gives us). So at the end of processing, we revert all the digits.

This does not work when rendering to a file, and had a very nasty side effect,
namely that it would corrupt memory because the rendering would revert a random
number of bytes *before* the scratchpad.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>