Make it safe to have GC pointers in C++ globals
There is no guarantee on the order of globals initialization in C++.
Trying to add a global pointer in the Input global led to a subtle crash because
the runtime::RT global was initialised after the Input global, causing the GC
safe pointer list to be reset _after_ Input had registered a global gcptr
(see next commit for the use case).

Since there is a singleton runtime, there is no harm in having the GCSafe list
be static-initialized to nullptr, which makes it robust to being used during
globals initialization.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>