Use a GC-safe pointer for the menu object
The menu object may move in memory when you set it from the command line,
e.g. if you type `VariablesMenu`. This led to a subtle memory corruption issue
when the state save was modified to include the current menu. The memory
corruption happened as follows:

1/ Execute a menu command, e.g. `VariablesMenu` from the command-line that was
   loaded from the saved state.

2/ This registered the object on the command-line as the menu object.

3/ The command-line was then discarded, and the menu object replaced with some
   random object.

Note that this fix only worked after applying the fix in the previous commit,
because the global GC pointer `Input::menuObject` was initialized _before_ the
`runtime::GCsafe` global, which led to the `menuObject` not being registered as
a safe pointer during garbage collection.

NOTE: This fix is not very good, because it forces the GC to keep the
command-line loaded from disk to stay around until the menu is changed. In an
upcoming commit, we will clone the menu object to avoid that issue.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>