Seriously reworked the rendering of decimal numbers
The existing code was really hard to read and maintain. I could not get FIX to
really work the way it's supposed to be, so restarted from the ground up.

While I was thinking about this, added the `SIG` mode for display, which
displays significant digits. I remember some other calculator having that, and
it's neat. It's like `STD`, but with a given precision.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>