Accept both decimal separators as input
Do not depend on the user preference to accept '.' or ',' for decimal
numbers.  This makes files on disk more portable, and the syntax less
ambiguous.  The shift of `,` to `;` for complex number or function
argument separators was always a bit dubious.

For complex numbers, we will use the 'i' notation anyways, i.e. `x 𝒊y`
or `x ∡y` for rectangular or polar, rather than the HP48's
mode-dependent `(x,y)` or `(x;y)`. Note that `𝒊` is a dedicated
character, codepoint 0x1D48A.

For algebraic function, the best syntax is still being under
consideration.  The problem is that `f(1,0,x)` can be either three
arguments, two of them being interegers, or two arguments, the first one
being `1,0` in European notation.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>