Attempt to build a QSPI image that is compatible with DM42 program
The idea is to keep the same symbols that the DM42 software uses in
a compatible location, to make it possible to alternatively load the
DM42 and DB48X program.

This seems to work relatively well, but there is a strange behavior
from the DM42 program that complains about being unable to open
the STATE directory of the calculator. It is possible that it is
confused by the presence of .48S files there.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>