Fix selection of arithmetic precision for decimal numbers
The implemented strategy was to select a type based on input types.
This causes issues if computing `1.0/3`, which is computed
using `decimal32`, and therefore outputs `0.3333333`.
Similarly, squaring `1.0e80` leads to infinity, since `1.0e80`
fits in a `decimal32` and that format cannot hold `1.0e160`.

Use the precision settting instead to determine a minimum
target type for computations. The current default of 30 selects
the `decimal128` type.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>