Binary computations in arbitrary bases
When using the a specific suffix, e.g. `#1234Ah`, you force the base of the
object to be hexadecimal. It does not change even if you change it on display.

When not using a specific suffix, e.g. `#1234A`, the current base setting is
used, and the display changes when the `Base` setting is changed.

Fixes #18

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>