Refactor the object model
Instead of going through an indirection and a switch for every single
command, go through a single indirection through a big table of
operations, with shared function pointers deduced from normal object
inheritance rules.

This reduces the overall size of the generated binary by about
8K (from 548468 down to 540588), at the expense of an increase in the
QSPI size (the dispatch table is larger), and makes the execution
of the NQueens program faster (850ms instead of 990ms).

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>