Create a C++ interface for RPL arithmetic
In the implementation of `complex` or `matrix` objects, we will
need to easily write code that performs RPL arithmetic.

In addition, reordered `x` and `y` arguments in most of the
existing arithmetic code, which was hard to read by `x` was the X register
on the RPL stack (level 0) and Y was the Y register (level 1),
so all operations were written `y - x` instead of `x - y`.
This made the code hard to read in particular when connected with
operations on bignum or fractions that took arguments in the "right" order.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>