Implementation of for loop
The `for next` and `for step` loops were half implemented, but did
not declare the local variable. Now that we have the infrastructure
for local variables, it is possible to implement them properly.

Furthermore, the parsing of `Start..Next` and `Start..Step` would re-parse
the _entire_ body of the loop twice, once to match `next`, the second to
try with `step`. This was horrendously inefficient, all the more so because
this is some of the most complex parsing we have. Refactored the code so
that it parses once, and decides the type of the generated object depending
on the closing delimiter (`step` or `next`).

Fixes: #92

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>