Remove remaining occurences of static_type()
The new model for dynamic typing uses a `static_id` field.
When we had a `static_type()`, this meant we had no `static_id`,
and a dynamic cast to the object would fail, returning a `nullptr`.

This notably impacted the promotion of negative numbers to decimal.
A simple test case was `-12345 LN`, which would crash.

Probbaly fixes #101, but further testing is required.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>