Fix crash detected in the simulator running the test suite
In the `runtime::close_editor()` funtion, we were allocating
header bytes for the string being copied, as well as a trailing zero.
The trailing zero was not passed to the `available()` check,
which in a corner case, allowed us to allocate past the red zone.
Once this happens, `available()` keeps returning enough space,
since the size computations are unsigned.

This could probably happen on a real machine, not just in the simulator,
although the simulator was way more likely to hit it since it runs in 2K of
memory, thereby being close to the redzone a lot more often.

Fixes: #118

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>