Avoid freeing scratch buffer space after putting it in editor
The `edit()` function is a bit special, in that it "reallocates" the
scratch space directly into the editor, without needing to move or
copy anything. However, since now the renderer also accounts for what
it puts in the scratchpad, and attempts to clean it up when the renderer
is destroyed, we need to tell it that the memory should not be freed anymore.

This was harmless because there is defensive coding against freeing more
than what was allocated, and in that scenario, `Scratch` was always zero.
But some added instrumentation tripped on that case.

Fixes: #118

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>