Fix evaluation of "ABC"*3
The evaluation of non-numeric cases was broken because a C++ template
specialization was not visible at the point of instantiation. The code
had been moved to a header, where the specializations were not visible,
to avoid a linker error for the special case of atan2 and hypot (which
are in a different file). Add explicit specialization at the right spot
instead.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>