Refactor the C++ interface for arithmetic
Change the interface for the various arithmetic functions with the
following guiding principle:

1. Return an `algebraic_p` or similar when possible, and not an
   `algebraic_g`.  The `_g` version can be constructed on return if
   necessary, but since no garbage collection can ever occur on a
   return path, there is no value in constructing / destructing `gcp`
   instances.

2. However, for functions that "cascade" their result, return an
   `algebraic_g` because that's automatically converted to an
   `algebraic_r`, making it simpler to write expressions.

3. Pass input as `algebraic_r` or similar. This is a `const gcp<T> &`,
   which is lightweight to pass around, and is consistent with regular
   C++ semantics.

4. Create temporary local copies as `algebraic_g` only when useful.

Also refactored the `non_numeric` interface to return an `algebraic_p`
instead of a `bool`. Unlike the arithmetic promotion cases, there is
no in-place change of the value.

There are still some possible improvements, notably to make it
easier to call functions. Currently, you need `sq::run(x)`, which
is a bit verbose, but there is a namespace conflict with the
name of the structure.

Fixes: #131

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>