tests: Better synchronization with display thread
Better synchronization between the test suite and the display.
This accelerates the tests when there is no animation.

The root cause was that we were writing `lcd_update` at the end
of `refreshed()`, which would cause a slow-down if this was
called twice. Since this is called in `ready()`, which is
called by `expect()` or other tests, it is normal to call it
multiple times.

Fixes: #77

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>