test: Yet another pass at synchronization logic
With the new alpha-shifting logic, the test suite runs much faster
than before, and as a result, ran into many more synchronization issues.

Ensure that we wait for the stack to actually update before we
proceed with evaluating its content. This required a change in logic
where we expect the stack to actually change value, so some tests
have to be changed accordingly.

The net result is that the tests can now pass reliably at full speed.

Fixes: #77

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>