Display factorial as a postfix
This exposed a problem with menu rendering that is also fixed here, which
was a dependency on the first byte of a symbol being non-printable. It
turns out that we now had enough objects for that first byte to be ... `!`
which is not just printable, but led to weird bugs if you put `"!"` as a
menu label (or a `!` character as the first byte).

Fixes: #144

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>