tests: Synchronize with RPL thread in optimized redraw mode
The redraw optimization broke the test suite because the display would
no longer redraw as often. Add an explicit synchronization mechanism
between the RPL thread and the test thread where the test thread sends
a `KEYSYNC` pseudo-key, and the RPL thread forces a full display refresh
when that happens (mostly to cause a stack update at a known time),
and notifies the test thread when the display update completed.
That way, the test thread only reads the stack state _after_ the
required refresh.

Fixes: #155

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>