parser: Fix parsing of `Invert(X)`
There is a `disambiguate` flag to refuse parsing fancy names
`x²` and `x³` (which should be parsed as `sq(X)` and `cubed(x)`
respectively. This flag was incorrectly applied to `Invert`, which
was as a result treated as symbol.

This led to exploring why `ID_inv` should not have a fancy form of
`x⁻¹` for consistency with the rest... And then why the insert
for the postfix commands was totally bogus.

This commit adds a new mode for insertion, which is `POSTFIX`, and
that deals correctly with these postfix forms.

Fixes: #163

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>