Implement `GET` function
Implement the `GET` function on list, arrays and text.

Getting a character from a text value is a DB48X extension, which I
thought was in newRPL based on documentation, but experiments show
that does not work at the moment.

Getting an element from a two-dimensional array is not going to work
the same as for RPL, and I wonder if I _want_ to make it work the
same. In traditional RPL, `[ [1 2 3] [4 5 6] ] 1 GET` gives you
`1`. In the DB48X implementation, it gives `[1 2 3]`. I think that
being able to extract vectors from a matrix is useful, and the ability
to extract an individual element using a list index should address
that problem.

Not yet implemented:
- `GET` with a list index for matrices
- `GET` with a name (dereferencing the name)

Fixes: #173

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>