det: Implement determinant on square matrices
General algorithm to compute the determinant

Compute the determinant by diagonalizing the matrix.
All computations are done in a way that will be conducive to getting
the correct result with integers or fractions. This is a step towards
implementing matrix invert and division.

The algorithm used is somewhat inspired from [this source][det], but
adapted for the RPL implementation in DB48X. Specifically, the matrix
is "exploded" to the stack by the `is_matrix()` function, which
returns the size of the matrix and puts all elements on the stack if
successful. The stack is then used as an array of value during the
various manipulations. Note that in the case of `det`, we can use a
single array, but that this technique can be extended to have multiple
arrays, which will be necessary for the division.

[det]:v https://www.geeksforgeeks.org/determinant-of-a-matrix

Fixes: #187

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>