Fix forced dereference in gcptr operator*
We need to return an _address_ because our objects _cannot_ be dereferenced.
This manifested as a bad `this` pointer given to `begin()` when building
an iterator in `for (object_p obj : *from)` where `from` was an `equation_r`.
Instead of passing the address in the `.safe` field, the compiler was
dereferencing it.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>