Add transient alpha mode
After a lot of experimentation, it turns out that it's not really
possible to use the shift key as originally planned, because it is not
reliably visible when two keys are simultaneously visible. It can only
be combined with key N and O (and maybe a few others), and these
combinations are clearly treated specially (they give some magic
keycode, e.g. screen snapshot with Shift-DISP on the DM42).

It turns out only three keys seem somewhat reliable in combination
with alphabetic keys: ON, UP and DOWN. On works with _almost_ all
keys, but apparently B or C is missing (forgot which one). Up and Down
work with all alphabetic keys. This turns out to be a good thing,
because it gives us an additional option that was not originally
planned: quick access to lowercases.

The behavior implemented in this commit is therefore the following:
- "Up" held with any key gives an uppercase letter
- "Down" held with any key gives a lowercase letter.

A minor side effect is that up and down are now active on release and
not on key press. In practice, this is barely noticeable.

Fixes: #204

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>