Add `expand`, `collect` and `simplify` functions
Add the basic rules for `expand`, `collect` and `simplify`.

And to be honest, I am quite proud of the fact that the core code looks
_exactly_ like this (this is actual code):

```c++
    return rewrite((x+y)*z,     x*z+y*z,
                   x*(y+z),     x*y+x*z,
                   (x-y)*z,     x*z-y*z,
                   x*(y-z),     x*y-x*z);
```

A lot of template metaprogramming to turn that into transparent
building of _compile-time constants_ representing the RPL objects for
the equations, and passing that to the `rewrite` function that was
implemented earlier.

What is really neet about this is that the generated equations:

* Are automatically _shared_ if used multiple times, because they are
  represented by the same template. So the equations used in `collect`
  and those used in `expand` (which are the same in reverse) share the
  same memory.

* Automatically go into QSPI, not consuming precious flash space,
  because they are represented as initialized data arrays, and I
  requested that all C++ .rodata go in QSPI.

I intentionally did not support the historical names
(`expan` and `colct`). Our level of compatibility with the HP48 is
never going to be that good anyway.

The code as is works with `clang`, but runs into unsatisfied symbols
with g++ 10.3 due to what I believe is a g++ bug (that's what living
on the edge of template metaprogramming gives you):

```
/opt/homebrew/Cellar/arm-none-eabi-gcc/10.3-2021.07/gcc/bin/../lib/gcc/arm-none-eabi/10.3.1/../../../../arm-none-eabi/bin/ld: build/release/equation.o: in function `equation::collect() const':
equation.cc:(.text._ZNK8equation7collectEv+0xd0): undefined reference to `eq<(unsigned char)10, (unsigned char)1, (unsigned char)120, (unsigned char)10, (unsigned char)1, (unsigned char)122, (unsigned char)57, (unsigned char)10, (unsigned char)1, (unsigned char)121, (unsigned char)10, (unsigned char)1, (unsigned char)122, (unsigned char)57, (unsigned char)55>::object_data'
```

Fixes: #231

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>