rewrite: Add "unique" variable names
Beginning a rewrite variable name with `u`, `v` or `w` forces it to be
unique across all variables. This is useful to ensure that reordering
terms do not match "no-op".

For example, if you want to rewrite `A+B+A` as `A+A+B`, but do _not_
want this to apply to `A+A+A` where it does nothing, then you can
use the pattern `X+U+X`, where the `U` (for "unique") ensures that
there is no match in the `A+A+A` case where that would make `U`
identical to `X`.

Fixes: #244

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>