complex: More careful handling of angle units for polars
The polar representation of complex has an angle that is now always
now represented as a fraction of pi. When the change happened,
insufficient care was given to angle conversions that occur in the
functions being called (e.g. `sin` or `cos`) as part of the
implementation of `re` or `im`. As a result, many functions would give
an incorrect result on polar numbers, including simple ones like
`conj`.

In order to preserve accuracy across successive polar <-> rectangular
conversions, functions like `sin`, `cos`, `tan` or `atan2` now have
special cases to deal with exact results. This guarantees that
`cos(90)` will give exactly zero in degrees mode, not something really
close to zero, but not exactly zero due to rounding of 90 degrees to
radians before processing by the BID128 library.

Fixes: #251

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>