Implement the `bytes` command
The `bytes` command on HP calculators returns the size of the object
in the first level of the stack, and the CRC32 (hash) of the object
in the second level fo the stack.

The implememntation here returns the _complete_ representation of the
object in level 2. Also, the size is always an integral number of
bytes (the Saturn-based calculators were 4-bit, so byte sizes could be
half integer)

The lowest part of the number matches the result of the `type` command.

Fixes: #284

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>