Implement type and typename commands
Implement the `type` command that returns the type of an object as an
integer. No attempt is made to match the return value of `TYPE` on the
HP48. Maybe someday if this has any value.

Add the `typename` which is a portable version of the above, returning
the type name as text.

Fixes: #285

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>