Fix pointer issues with GC during constructor
The decimal128 constructor from fraction is special in that it may
trigger GC during construction, since it calls `fraction::numerator()`
and `fraction::denominator()`, which may create `bignum` objects.

This could cause the `runtime::make` function to return the pre-GC
pointer, which contained garbage. This in turn led to frequent
premature termination of the `Draw` plot loop.

A couple of additional places where a GC-safe pointer was needed
were found during inspection, notably in `real_promotion` for `bignum`
and `fraction` types, although they were not triggering this
specific bug.

Finally, comparison code was brought to the "new standard" used for
arithmetic passing `algebraic_r` instead of `algebraic_g` to minimize
the number of `gcp` constructors/destructors.

Fixes: #325

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>