Plot: Accept functions that use the independent variable from PPAR
When drawing an equation, it should be possible to use a form of equation that
works like on the HP calculators, e.g. `'sin(1000*X)*gamma(X)'` where `X` is the
name given in `PPAR`.

Howevever, I also like the current behavior of having a program that consumes an
argument from the stack and returns it on the stack. We should be able to
auto-detect the behaviour.

This commit supports both behaviours. The indepdendent variable is marked as
special for lookup. This seems relatively similar to what the HP48 does (at
least, on HP48, I see no global variable change during plotting).

Fixes: #352

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>