Implement source-code comments
This commit implements comments that are preserved in the source code.

A comment begins with `@` and ends with the end of line.
It acts as a no-operation during execution.
It is rendered beginning with a new-line.

This commit also adjusts various issues related to rendering of
newlines and indentation in loops or blocks.

Fixes: #94

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>