Implement Tag type
The tag type is used to add a label to any arbitrary object.
It is parsed as `:Label:object`, and renders the same, except on the
stack where it renders as `Label:object`.

Note that the HP48 renders a label with an extra space. The space is
apparently gone on the HP50.

Fixes: #21

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>