blitter: Fix off by one error when swapping coordinates
When we swap left and right coordinates, if the width is 384, we need
0 to turn to 383, not 384. Otherwise, we end up writing one pixel
beyond the actual data.

The most obvious manifestation is that since graphical rendering of
the stack was enabled, it was possible that the off-by-one pixel
update would end up overwriting the beginning of the text editor,
turning `B` into `C`, etc.

Fixes: #386

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>