stack: Switch back to text mode
The graphical stack rendering presents no benefit at the moment.
It consumes more memory, and is not any faster, because we don't
cache the results yet. Furthermore, it is not full-featured yet
either.

Since this causes a significant slow-down on very large objects,
let's just default to text mode for the moment.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>