graphics: Fix the semantics of the low-level blitops
The semantics of the blitops was incorrect: it was performing the operation
only with the colors, not with the source. This impacted `GXor`, `GOr` and
`GAnd`.

However, it's still necessary to keep the color argument, because there is this
weird inversion of colors between platforms (e.g. which bits represent black or
white). We also need a pattern to indicate inversion.

Fixes: #399

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>