editor: Make the editor fast when `VariablesMenu` is active.
The mechanism for updating the `VariablesMenu` was too aggressive.
It updated the whole series of menus after every single key.
The same thing was true with the catalog.

Change to a mechanism where updates are only made when necessary.
For example, the `VariablesMenu` only updates when a `Sto` or similar
operation changes the content of the directory.

Fixes: #404

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>