complex: Fix complex `ln` and power operator when not in radians
When the change in complex representation was made to use internal powers of pi,
as well as to support angular modes, an internal uses of the `atan` function to
compute `pi` was left there by mistake. That function became sensitive to angle
mode. As an indirect result, the computation of `arg()` and any of its
derivative uses, including `log` or the exponentiation operator, would give the
wrong result when the angle mode was not radians.

This was not seen in the regression suite, which always runs this part of the
computation checks in radians...

Fixes: #452

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>