equation: Parse `(sin x)²+(cos x)²` correctly
The problem is that when parsing parentheses, we would not keep
parsing to grab the `²`, so we processed `+` before looking again.
As a result, we parsed as `((sin x)²+cos x)²`.

Another issue was found in something like `2*(3-1)^4`, because
the opening parenthese would cause us to pick the wrong precedence
for `(3-1)^4` subexpression. This was fixed as well.

With this fix, we now parse the "Mach number computation" from page 106
of the [HP-67 user's manual correctly][hp67], and output:

`'√(5×(((((1+0.2×(350÷661.5)²)↑3.5-1)×(1-6.875⁳-6×25500)↑-5.2656)+1)↑0.286-1))'

Fixes: #427

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>

[hp67]: https://literature.hpcalc.org/community/hp67-oh-en.pdf