ui: Do not update `LastArg` except for command line
Saving `LastArg` is a relatively expensive operation, and while it is
extremely useful for interactive operations, is is much more rarely
used in programs.

The HP48 controls the saving of `LastArg` with flag -55, which is a
global setting, also shared by RPL when executing user programs. A
side effect is that if you run `'3*X-5' 'X' 0 ROOT`, the `LASTARG`
command gives a useless result (generally 0). An interactive user
might expect to get the parameters to the `ROOT` command, not the last
parameters used by some internal `ROOT` operation.

The code implemented in this change decouples that into two distinct
settings, one that controls the saving of `LastArg` for interactive
commands, i.e. either direct commands or for operations on the command
line, and another that controls the saving of `LastArg` while running
programs. The first one is enabled by default, the second one is
disabled by default.

An additional setting was defined to save the last stack.

Controling the saving of `LastArg` is important both for performance
and functionality.

Fixes: #511

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>