fonts: Reimplement the font cache to be glyph-based
Reimplement the font cache so that it caches individual font/glyph
combinations, which proves much more efficient than the range-based
approach that had been taken earlier.

Experiments have shown that with a 128-entries glyph cache, on typical
use such as displaying numbers with scientific notation, walking
through the help file or editing the version text, the cache hits over
99% of the time (very often 0% misses while editing or doing
computations), and the average walk throught the cache is in the low
tens (typically between 11 and 15).

With this fix, DB48X on battery feels at least is as snappy as the
DM42 stock firmware doing arithmetic.

Fixes: #526

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>