fraction: Fix 32-bit truncation on calculator
The `fraction::numerator()` and `fraction::denominator()` functions
incorrectly used `size_t` where `ularge` was intended. This resulted
in a truncation of the value during rendering of the fraction, as well
as in arithmetic operations that isolate numerator and denominator. So
the fraction was correct internally, but displayed and computed
incorrectly.

Fixes: #539

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>