gc: Detect and move menu labels
Menu labels can point to global variables. They need to be garbage
collected correctly if the variable is moved in global storage.

This was a non-problem because the variables menu gets updated
whenever a variable name could possibly be moved. However, if we want
to be able to put the name of units loaded from a file in the units
menu, it is important that this also would identify dynamic objects to
keep.

Fixes: #548

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>