runtime: Rework program run loop and call stack management
The HP48 keeps its call stack in normal garbage-collected memory.
This makes it possible for a program like the following to run until
it exhausts available memory:

```
<< 1 + DUP 1 DISP PROG >> 'PROG' STO
1 PROG
```

The earlier DB48X implementation used the C call stack, which limited
the above program to about 40 iterations on the DM42 before crashing.
Such a limitation is likely to be much too restrictive for complicated
recursive algorithms.

There was an early attempt at writing code for a call stack, but that
code had not been maintained. The code was updated in this commit so
that it correctly adjusts all the required runtime pointers when
entering a call.

The main program run loop was rewritten to avoid consuming C++ stack
for RPL recursion. Instead, it uses the above RPL stack for that. We
can still run out of memory, but this can now be processed as a
regular RPL error instead of crashing due to C++ stack exhaustion.

Fixes: #537

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>