runtime: Remove the `execute` callback
Defer program and expression execution to the `program` class.
This allows us to remove `execute()`, which applies to every object,
with `run()`, which only exists for `program`. There is also a
`program::run()` static member that dynamic checks if we should run
`evaluate()` or if we can `run()`.

Also reimplement `while`, `until`, `if`, `ift`and `ifte` using a
deferred conditional so as to avoid C++ stack recursion.

This allows us to have really good behaviour on tail recursion, see
the `Collatz` and `CBench` examples in the `Demo.48S` file. For this
particular case, DB48X on DM42 is 25x faster than the HP50G, which
becomes slower as the recursion depth increases.

Fixes: #537

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>