recall: Evaluate equation independent and dependent variables
The recent changes in `Sto` and `Recall` changed the type of `name`
to `object_p`, which causes the wrong (case-dependent) version of
`is_same_as` to be called (`object::is_same_as` instead of
`symbol::is_same_as`).

Restrict the lookup of `expression::independent` and
`expression::dependent` to the name case, with a name type.

Fixes: #583

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>