plot: Implement bar and scatter plot
Implement a bar plot and a scatter plot function.

The bar plot differs from the HP calculators in that it can take pair
of values, which are interpreted as high/low ranges.

The scatter plot differs from the HP calculators in that it connects
data points by default. This can be disabled with `NoCurveFilling`

Fixes: #577
Fixes: #579

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>