graphics: Add `gray` and `rgb` pattern commands
The `GrayPattern` and `RGBCommand` generate a grayscale or colored
pattern for use by `Foreground` or `Background`. They take an input
between 0 and 1. For example, `0.25 Gray Foreground` generates a 25%
gray pattern, and `1 1 0 RGB Background` geneates a yellow background.

Existing calculators running DB48X currently only have black and white
screens. The commands generate a pattern that simulates the
grayscale. The pattern is always 64-bit, representing an 8x8 pixel
grid.

However, the blitter code has been tested on simulators for the HP50G
and the HP Prime. The HP50G has a 4-bit grayscale screen, and the HP
Prime has a 16 bit per pixel 5-5-6 RGB screen. On such platforms, the
command will generate a real grayscale and a filled color pattern.
The 64-bit in that case represent either a 4x4 grayscale pattern, or a
2x2 color pattern. Using the `GrayPattern` and `RGBPattern` commands
will ensure your code remains portable if/when DB48X is ported to
these platforms.

Fixes: #617

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>